Move identity submodule into components
