+++ src/App.tsx
@@ -11,12 +11,12 @@ import CliInstructions from "./components/transferData/CliInstructions";
 import TemplatesPage from "./components/templates/TemplatesPage";
 import PrivacyAndSecurityPage from "./components/privacyAndSecurity/PrivacyAndSecurityPage";
 import UserAccountPage from "./components/userAccount/UserAccountPage";
-import Register from "./identity/Register";
-import Unactivated from "./identity/Unactivated";
-import history from "./identity/History";
+import Register from "./components/identity/Register";
+import Unactivated from "./components/identity/Unactivated";
+import history from "./components/identity/History";
 import AssayInstructions from "./components/transferData/AssayInstructions";
-import AuthProvider from "./identity/AuthProvider";
-import UserProvider from "./identity/UserProvider";
+import AuthProvider from "./components/identity/AuthProvider";
+import UserProvider from "./components/identity/UserProvider";
 import ErrorGuard from "./components/errors/ErrorGuard";
 import InfoProvider from "./components/info/InfoProvider";
 import { createMuiTheme, MuiThemeProvider } from "@material-ui/core";
