+++ src/components/identity/Register.tsx
@@ -12,8 +12,8 @@ import {
     CircularProgress
 } from "@material-ui/core";
 import "./Register.css";
-import { ORGANIZATION_NAME_MAP } from "../util/constants";
-import { createUser } from "../api/api";
+import { ORGANIZATION_NAME_MAP } from "../../util/constants";
+import { createUser } from "../../api/api";
 import { AuthContext } from "./AuthProvider";
 import history from "./History";
